refactor(consumers): update to pure esm
