+++ packages/consumers/package.json
@@ -1,9 +1,13 @@
 {
   "name": "@drupal-kit/consumers",
   "version": "0.3.1",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "dist/index.d.ts",
+      "import": "dist/index.mjs"
+    }
+  },
   "dependencies": {
     "@drupal-kit/core": "workspace:0.3.1",
     "@wunderwerk/ts-results": "^0.1.1"
@@ -12,23 +16,32 @@
     "@drupal-kit/config-typescript": "workspace:0.3.1",
     "@drupal-kit/eslint-config": "workspace:0.3.1",
     "@drupal-kit/types": "workspace:0.3.1",
-    "@rollup/plugin-typescript": "^11.0.0",
-    "@types/jest": "^29.4.1",
-    "esbuild": "^0.17.11",
-    "jest": "^29.5.0",
-    "jest-fetch-mock": "^3.0.3",
-    "rollup": "^3.19.1",
-    "rollup-plugin-dts": "^5.2.0",
+    "@rollup/plugin-typescript": "^11.1.0",
+    "ava": "^5.2.0",
+    "esbuild": "^0.17.16",
+    "msw": "^1.2.1",
+    "rollup": "^3.20.2",
+    "rollup-plugin-dts": "^5.3.0",
     "rollup-plugin-esbuild": "^5.0.0",
-    "ts-jest": "^29.0.5",
-    "tslib": "^2.5.0",
-    "typescript": "4.9.5"
+    "ts-node": "^10.9.1",
+    "typescript": "5.0.4"
   },
   "scripts": {
     "build": "rollup -c",
     "lint": "eslint --ext .ts,.tsx src --max-warnings 0",
-    "test": "jest",
-    "test:watch": "jest --watch",
+    "test": "NODE_NO_WARNINGS=1 ava",
     "typecheck": "tsc --project ./tsconfig.json --noEmit"
+  },
+  "ava": {
+    "files": [
+      "tests/**/*.test.ts"
+    ],
+    "extensions": {
+      "ts": "module"
+    },
+    "nodeArguments": [
+      "--loader",
+      "ts-node/esm"
+    ]
   }
 }
