+++ packages/consumers/rollup.config.mjs
@@ -1,24 +1,18 @@
 import dts from "rollup-plugin-dts";
 import esbuild from "rollup-plugin-esbuild";
 
-import packageManifest from "./package.json" assert { type: "json" };
-const name = packageManifest.main.replace(/\.js$/, "");
-
 const bundle = (config) => ({
   ...config,
   input: "src/index.ts",
   external: (id) => !/^[./]/.test(id),
 });
 
+const name = "dist/index";
+
 export default [
   bundle({
     plugins: [esbuild()],
     output: [
-      {
-        file: `${name}.js`,
-        format: "cjs",
-        sourcemap: true,
-      },
       {
         file: `${name}.mjs`,
         format: "es",
